Add note about parameter server that is potentially breaking object initialization.
When starting launching the node multiple times without restarting the
roscore, the trajectory_file parameter may still be defined on the
parameter server even if it was not defined in the launch file that
started the node. The error message now clarifies that.

Also applying mrt formatting introduced one or two tiny changes.